+++ backend/database/migrations/2023_05_29_073808_create_payments_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->bigInteger('payment_method_id')->unsigned();
             $table->bigInteger('payment_status_id')->unsigned();
             $table->date('payment_date')->nullable();
-            $table->float('total_price')->nullable();
+            $table->decimal('total_price',12,2)->nullable();
             $table->string('note')->nullable();
             $table->bigInteger('create_user')->nullable();
             $table->bigInteger('modified_user')->nullable();
